backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
boto3==1.35.22
botocore==1.35.22
certifi==2024.7.4
charset-normalizer==3.3.2
cohere==5.9.4
colorama==0.4.6
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fastavro==1.9.7
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.0
idna==3.7
jmespath==1.0.1
Markdown==3.6
packaging==24.1
parameterized==0.9.0
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
